+++ build_runner/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.2.0-dev
+
+- Support global 'runs_before' configuration for builders. This allows users
+  to specify the ordering of builders in their build.
+
 ## 2.1.11
 
 - Stop reading or requiring `.packages` files.